+++ front/package.json
@@ -20,6 +20,7 @@
     "react-router-dom": "^6.14.2",
     "react-scripts": "^5.0.1",
     "react-slick": "^0.29.0",
+    "react-toastify": "^10.0.5",
     "react-typist": "^2.0.5",
     "recoil": "^0.7.7",
     "slick-carousel": "^1.8.1",
