--- front/src/App.css
@@ -1,11 +1,0 @@
-/*@font-face {
-  font-family: "Pretendard Variable";
-  font-weight: 45 920;
-  font-style: normal;
-  font-display: swap;
-  src: url("./assets/fonts/PretendardVariable.woff2") format("woff2-variations");
-}*/
-
-/*body {
-  font-family: "Pretendard-Regular";
-}*/
