+++ front/src/component/meeting/MeetingDetail/MeetingApplyReasonModal.jsx
@@ -2,13 +2,29 @@ import TextArea from "../../textArea/TextArea";
 import { meetingApplyReasonState } from "../../../recoil/contentState";
 import Button from "../../button/Button";
 import useInputGlobal from "../../../customhook/useInputGlobal";
+import { useRecoilValue } from "recoil";
+import { userState } from "../../../recoil/userState";
+import {
+  userNameState,
+  userPhoneNumberState,
+  emailState,
+  userBirthdayState,
+  userGenderState,
+} from "../../../recoil/authState";
 
 import { GoX } from "react-icons/go";
+import { Toast } from "../../toast/Toast";
 
 export default function MeetingApplyReasonModal({
   meetingApplyReasonModal,
   setMeetingApplyReasonModal,
 }) {
+  const userIn = useRecoilValue(userState);
+  const userName = useRecoilValue(userNameState);
+  const userPhoneNumber = useRecoilValue(userPhoneNumberState);
+  const userEmail = useRecoilValue(emailState);
+  const userBirthday = useRecoilValue(userBirthdayState);
+  const userGender = useRecoilValue(userGenderState);
   const [meetingApplyReasonInput, onMeetingApplyReasonChange] = useInputGlobal(
     meetingApplyReasonState
   );
@@ -21,9 +37,7 @@ export default function MeetingApplyReasonModal({
     >
       <div className="w-full flex items-center justify-end">
         <button
-          onClick={(e) => {
-            e.preventDefault();
-
+          onClick={() => {
             setMeetingApplyReasonModal(false);
           }}
           className="w-8 h-8 bg-[#ffc35c] rounded-full flex items-center justify-center text-white text-[1.4rem] hover:[transform:rotateY(360deg)] transition-all duration-700"
@@ -56,6 +70,21 @@ export default function MeetingApplyReasonModal({
         <Button
           type="button"
           buttonText="신청하기"
+          onClick={() => {
+            if (userIn) {
+              if (
+                !userName ||
+                !userPhoneNumber ||
+                !userEmail ||
+                !userBirthday ||
+                !userGender
+              ) {
+                Toast("🍒🍓먼저 마이페이지에서 개인정보를 채워주세요!");
+              }
+            } else {
+              Toast("🍇🍋먼저 로그인을 진행해주세요!");
+            }
+          }}
           disabled={meetingApplyReasonInput.length === 0}
           basicStyle="h-auto mt-2"
         />
